+++ CHANGELOG.md
@@ -9,7 +9,9 @@
 
 * Implement class and singleton-class methods in Swift.
 * Define classes and modules from Swift.
-* Add `RbObject` to module mix-in functions.
+* Add module mix-in functions to `RbObject`.
+* Bind Ruby objects and methods directly to Swift objects and methods.
+* Add throwing conversion as alternative to optional initializer.
 
 ##### Bug Fixes